First Commit
Created the SudokuSolver class which may be inherited later when
building a graphical representation. Various helper functions to perform
logic built; whether or not they will actually be necessary is yet to be
seen. I started out approaching the problem from one angle, and then
figured out that there was a more suitable way to do it.

The way I plan to build this program is outlined in the Algorithm file.
Essentially, every number 1-9 has its own grid with 0's in all the
cells. For every number added to the puzzle grid, the cell it is added
at becomes a 1 for each number grid (signifying that it is no longer a
usable space). In addition, the number grid corresponding to the number
itself also has 1's entered in the entire row, column, and section
corresponding to the number (since no number can be repeated in the same
row, column, section).

This algorithm, though not completed in the code yet, should work for
simple sudoku puzzles; cases where one must consider columns/rows on
their own or must try multiple cases are not yet factored into the
algorithm.