In order to support recursive calls, modified all method parameters to include grid[][] variables as opposed to using class fields (which made recursion impossible). This version uses a simpler algorithm which simply keeps track of how many possible values can go in a cell. It fills all the cells where only 1 value is possible, then finds the cells with the next least and tries multiple cases.
I will have to implement something to filter through all the failed
cases that get returned, but for now this works so I wanted to commit.

Background for this commit: I took the free time I had over the past
week and a half to contemplate the algorithm I initially created. Then,
when I sat down to code today I realized that it's pretty difficult to
create an algorithm that works without actually sitting down and trying
to code it. As a result, I have started with a simpler algorithm which
is less memory intensive. It works with one test case so far, so before
I continue testing I want to commit this change.