Moved the majority of the functionality over to a Grid class. This makes a lot more sense in terms of OOP and making a larger program more easily possible.
In the Grid class exist several functions previously in the SudokuSolver
class; it really makes a lot more sense this way. SudokuSolver
refactored to use the Grid class methods.

Additionally, added the ability to instantiate a Grid using one long
string of values or through a byte[][] array. This will make it easier
to use once the program goes beyond command line input.